+++ settings.gradle
@@ -7,7 +7,7 @@ include ':vtm-app'
 include ':vtm-desktop'
 include ':vtm-extras'
 include ':vtm-gdx'
-include ':vtm-gdx-models'
+include ':vtm-gdx-poi3d'
 include ':vtm-http'
 include ':vtm-ios'
 include ':vtm-ios-example'
