+++ vtm-gdx-poi3d/build.gradle
