+++ vtm-gdx-poi3d/src/VtmGdxPoi3D.gwt.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?><!DOCTYPE module PUBLIC
         "-//Google Inc.//DTD Google Web Toolkit trunk//EN"
         "http://google-web-toolkit.googlecode.com/svn/trunk/distro-source/core/src/gwt-module.dtd">
-<module rename-to="VtmGdxModels">
-    <source path="org/oscim/model/poi3d" />
+<module rename-to="VtmGdxPoi3D">
+    <source path="org/oscim/gdx/poi3d" />
 </module>
