+++ vtm-gdx-poi3d/src/org/oscim/gdx/poi3d/GdxModelLayer.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU Lesser General Public License along with
  * this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.oscim.model.poi3d;
+package org.oscim.gdx.poi3d;
 
 import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.g3d.Model;
