+++ vtm-gdx-poi3d/src/org/oscim/gdx/poi3d/ModelPosition.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU Lesser General Public License along with
  * this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.oscim.model.poi3d;
+package org.oscim.gdx.poi3d;
 
 import org.oscim.core.MercatorProjection;
