+++ vtm-web-app/src/org/oscim/web/VtmWebApp.gwt.xml
@@ -3,7 +3,7 @@
 
     <entry-point class="org.oscim.web.client.GwtLauncher" />
     <inherits name="org.oscim.gdx.VtmWeb" />
-    <inherits name="VtmGdxModels" />
+    <inherits name="VtmGdxPoi3D" />
 
     <set-property name='gwt.logging.enabled' value='TRUE' />
     <set-property name='gwt.logging.consoleHandler' value='ENABLED' />
